m3/requirements.txt
torch>=1.5.0
torchvision>=0.6.0


opencv-python>=4.5.0
pillow>=8.0.0


numpy>=1.19.0
scipy>=1.5.0

pandas>=1.2.0

matplotlib>=3.3.0

tqdm>=4.50.0

fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5

supabase>=2.0.0

python-dotenv>=0.19.0

requests>=2.25.0
